Add files via upload
xercise 1:

Given an array of numbers, sort ascending odd numbers but leave even numbers in their original place in the array.  Zero does not need to move. If the array is empty, return an empty array.


Examples:

[5, 3, 2, 8, 1, 4] should return [1, 3, 2, 8, 5, 4]

[6, 3, 2, 7, 5, 0] should return [6, 3, 2, 5, 7, 0]

Exercise 2:

Given a word, return the middle character or characters of the word.  If the word is odd in length, return a string of the middle character.  If the word is even, return a string of the middle two characters. If the string is empty, return an empty string.


Examples:

“dog” should return “o”

“Fishing” should return “h”

“have” should return “av”

“Middle” should return “dd”

“C” should return “C”